requirements.txt
Django==1.8
psycopg2==2.5.3
djangorestframework==3.3.2